[REFACTOR][I] Move session context factory to intellij package
Moves SarosIntellijSessionContextFactory to the package
'intellij.context'. This was done as the class contains Saros/I specific
dependencies, meaning it can not be moved to the core. For this reason,
it should not remain in the 'core' package.
